+++ CHANGELOG.md
@@ -25,6 +25,7 @@ CHANGELOG for Sulu
     * FEATURE     #2799 [MediaBundle]         Implemented tiles view for navigating through collections
     * FEATURE     #2715 [ContentBundle]       Implemented teaser content-type
     * ENHANCEMENT #2846 [ContentBundle]       Added teaser-attributes
+    * FEATURE     #2768 [ContentBundle]       Implemented teaser-edit for content-type
     * FEATURE     #2765 [MediaBundle]         Implemented new masonry-design for media
     * ENHANCEMENT #2743 [CoreBundle]          Remove symfony deprecations and don't allow them anymore
     * BUGFIX      #2810 [ContentBundle]       Add missing translation of Content navigation tab